requirements.txt
clamd
MySQL-python
python-ldap
spamc
ntplib
